+++ src/game/Object/Player.cpp
@@ -5104,55 +5104,6 @@ void Player::SendExplorationExperience(uint32 Area, uint32 Experience)
     GetSession()->SendPacket(&data);
 }
 
-/*********************************************************/
-/***              Update timers                        ***/
-/*********************************************************/
-
-/// checks the 15 afk reports per 5 minutes limit
-void Player::UpdateAfkReport(time_t currTime)
-{
-    if (m_bgData.bgAfkReportedTimer <= currTime)
-    {
-        m_bgData.bgAfkReportedCount = 0;
-        m_bgData.bgAfkReportedTimer = currTime + 5 * MINUTE;
-    }
-}
-
-void Player::UpdateContestedPvP(uint32 diff)
-{
-    if (!m_contestedPvPTimer || IsInCombat())
-    {
-        return;
-    }
-    if (m_contestedPvPTimer <= diff)
-    {
-        ResetContestedPvP();
-    }
-    else
-    {
-        m_contestedPvPTimer -= diff;
-    }
-}
-
-/**
- * @brief Updates and clears the player's PvP flag when the timeout expires.
- *
- * @param currTime The current server time.
- */
-void Player::UpdatePvPFlag(time_t currTime)
-{
-    if (!IsPvP())
-    {
-        return;
-    }
-    if (pvpInfo.endTimer == 0 || currTime < (pvpInfo.endTimer + 300))
-    {
-        return;
-    }
-
-    UpdatePvP(false);
-}
-
 /**
  * @brief Enables or disables the free-for-all PvP player flag.
  *
@@ -5168,35 +5119,6 @@ void Player::SetFFAPvP(bool state)
     {
         RemoveFlag(PLAYER_FLAGS, PLAYER_FLAGS_FFA_PVP);
     }
-}
-
-/**
- * @brief Starts an active duel once the duel countdown completes.
- *
- * @param currTime The current server time.
- */
-void Player::UpdateDuelFlag(time_t currTime)
-{
-    if (!duel || duel->startTimer == 0 || currTime < duel->startTimer + 3)
-    {
-        return;
-    }
-
-    // Used by Eluna
-#ifdef ENABLE_ELUNA
-    if (Eluna* e = GetEluna())
-    {
-        e->OnDuelStart(this, duel->opponent);
-    }
-#endif /* ENABLE_ELUNA */
-
-    SetUInt32Value(PLAYER_DUEL_TEAM, 1);
-    duel->opponent->SetUInt32Value(PLAYER_DUEL_TEAM, 2);
-
-    duel->startTimer = 0;
-    duel->startTime  = currTime;
-    duel->opponent->duel->startTimer = 0;
-    duel->opponent->duel->startTime  = currTime;
 }
 
 /**
@@ -5933,229 +5855,6 @@ void Player::UpdateHomebindTime(uint32 time)
         data << uint32(ERR_RAID_GROUP_NONE);                // error used only when timer = 0
         GetSession()->SendPacket(&data);
         DEBUG_LOG("PLAYER: Player '%s' (GUID: %u) will be teleported to homebind in 60 seconds", GetName(), GetGUIDLow());
-    }
-}
-
-/**
- * @brief Sets or clears the player's PvP state with timeout-aware handling.
- *
- * @param state True to enable PvP; false to disable it.
- * @param ovrride True to bypass the delayed PvP timeout behavior.
- */
-void Player::UpdatePvP(bool state, bool ovrride)
-{
-    if (!state || ovrride)
-    {
-        SetPvP(state);
-        pvpInfo.endTimer = 0;
-    }
-    else
-    {
-        if (pvpInfo.endTimer != 0)
-        {
-            pvpInfo.endTimer = time(NULL);
-        }
-        else
-        {
-            SetPvP(state);
-        }
-    }
-}
-
-bool Player::EnchantmentFitsRequirements(uint32 enchantmentcondition, int8 slot)
-{
-    if (!enchantmentcondition)
-    {
-        return true;
-    }
-
-    SpellItemEnchantmentConditionEntry const* Condition = sSpellItemEnchantmentConditionStore.LookupEntry(enchantmentcondition);
-
-    if (!Condition)
-    {
-        return true;
-    }
-
-    uint8 curcount[4] = {0, 0, 0, 0};
-
-    // counting current equipped gem colors
-    for (uint8 i = EQUIPMENT_SLOT_START; i < EQUIPMENT_SLOT_END; ++i)
-    {
-        if (i == slot)
-        {
-            continue;
-        }
-        Item* pItem2 = GetItemByPos(INVENTORY_SLOT_BAG_0, i);
-        if (pItem2 && !pItem2->IsBroken() && pItem2->GetProto()->Socket[0].Color)
-        {
-            for (uint32 enchant_slot = SOCK_ENCHANTMENT_SLOT; enchant_slot < SOCK_ENCHANTMENT_SLOT + 3; ++enchant_slot)
-            {
-                uint32 enchant_id = pItem2->GetEnchantmentId(EnchantmentSlot(enchant_slot));
-                if (!enchant_id)
-                {
-                    continue;
-                }
-
-                SpellItemEnchantmentEntry const* enchantEntry = sSpellItemEnchantmentStore.LookupEntry(enchant_id);
-                if (!enchantEntry)
-                {
-                    continue;
-                }
-
-                uint32 gemid = enchantEntry->GemID;
-                if (!gemid)
-                {
-                    continue;
-                }
-
-                ItemPrototype const* gemProto = sItemStorage.LookupEntry<ItemPrototype>(gemid);
-                if (!gemProto)
-                {
-                    continue;
-                }
-
-                GemPropertiesEntry const* gemProperty = sGemPropertiesStore.LookupEntry(gemProto->GemProperties);
-                if (!gemProperty)
-                {
-                    continue;
-                }
-
-                uint8 GemColor = gemProperty->color;
-
-                for (uint8 b = 0, tmpcolormask = 1; b < 4; ++b, tmpcolormask <<= 1)
-                {
-                    if (tmpcolormask & GemColor)
-                    {
-                        ++curcount[b];
-                    }
-                }
-            }
-        }
-    }
-
-    bool activate = true;
-
-    for (int i = 0; i < 5; ++i)
-    {
-        if (!Condition->Color[i])
-        {
-            continue;
-        }
-
-        uint32 _cur_gem = curcount[Condition->Color[i] - 1];
-
-        // if have <CompareColor> use them as count, else use <value> from Condition
-        uint32 _cmp_gem = Condition->CompareColor[i] ? curcount[Condition->CompareColor[i] - 1] : Condition->Value[i];
-
-        switch (Condition->Comparator[i])
-        {
-            case 2:                                         // requires less <color> than (<value> || <comparecolor>) gems
-                activate &= (_cur_gem < _cmp_gem) ? true : false;
-                break;
-            case 3:                                         // requires more <color> than (<value> || <comparecolor>) gems
-                activate &= (_cur_gem > _cmp_gem) ? true : false;
-                break;
-            case 5:                                         // requires at least <color> than (<value> || <comparecolor>) gems
-                activate &= (_cur_gem >= _cmp_gem) ? true : false;
-                break;
-        }
-    }
-
-    DEBUG_LOG("Checking Condition %u, there are %u Meta Gems, %u Red Gems, %u Yellow Gems and %u Blue Gems, Activate:%s", enchantmentcondition, curcount[0], curcount[1], curcount[2], curcount[3], activate ? "yes" : "no");
-
-    return activate;
-}
-
-void Player::CorrectMetaGemEnchants(uint8 exceptslot, bool apply)
-{
-    // cycle all equipped items
-    for (uint32 slot = EQUIPMENT_SLOT_START; slot < EQUIPMENT_SLOT_END; ++slot)
-    {
-        // enchants for the slot being socketed are handled by Player::ApplyItemMods
-        if (slot == exceptslot)
-        {
-            continue;
-        }
-
-        Item* pItem = GetItemByPos(INVENTORY_SLOT_BAG_0, slot);
-
-        if (!pItem || !pItem->GetProto()->Socket[0].Color)
-        {
-            continue;
-        }
-
-        for (uint32 enchant_slot = SOCK_ENCHANTMENT_SLOT; enchant_slot < SOCK_ENCHANTMENT_SLOT + 3; ++enchant_slot)
-        {
-            uint32 enchant_id = pItem->GetEnchantmentId(EnchantmentSlot(enchant_slot));
-            if (!enchant_id)
-            {
-                continue;
-            }
-
-            SpellItemEnchantmentEntry const* enchantEntry = sSpellItemEnchantmentStore.LookupEntry(enchant_id);
-            if (!enchantEntry)
-            {
-                continue;
-            }
-
-            uint32 condition = enchantEntry->EnchantmentCondition;
-            if (condition)
-            {
-                // was enchant active with/without item?
-                bool wasactive = EnchantmentFitsRequirements(condition, apply ? exceptslot : -1);
-                // should it now be?
-                if (wasactive != EnchantmentFitsRequirements(condition, apply ? -1 : exceptslot))
-                {
-                    // ignore item gem conditions
-                    // if state changed, (dis)apply enchant
-                    ApplyEnchantment(pItem, EnchantmentSlot(enchant_slot), !wasactive, true, true);
-                }
-            }
-        }
-    }
-}
-
-// if false -> then toggled off if was on| if true -> toggled on if was off AND meets requirements
-void Player::ToggleMetaGemsActive(uint8 exceptslot, bool apply)
-{
-    // cycle all equipped items
-    for (int slot = EQUIPMENT_SLOT_START; slot < EQUIPMENT_SLOT_END; ++slot)
-    {
-        // enchants for the slot being socketed are handled by WorldSession::HandleSocketOpcode(WorldPacket& recv_data)
-        if (slot == exceptslot)
-        {
-            continue;
-        }
-
-        Item* pItem = GetItemByPos(INVENTORY_SLOT_BAG_0, slot);
-
-        if (!pItem || !pItem->GetProto()->Socket[0].Color)  // if item has no sockets or no item is equipped go to next item
-        {
-            continue;
-        }
-
-        // cycle all (gem)enchants
-        for (uint32 enchant_slot = SOCK_ENCHANTMENT_SLOT; enchant_slot < SOCK_ENCHANTMENT_SLOT + 3; ++enchant_slot)
-        {
-            uint32 enchant_id = pItem->GetEnchantmentId(EnchantmentSlot(enchant_slot));
-            if (!enchant_id)                                // if no enchant go to next enchant(slot)
-            {
-                continue;
-            }
-
-            SpellItemEnchantmentEntry const* enchantEntry = sSpellItemEnchantmentStore.LookupEntry(enchant_id);
-            if (!enchantEntry)
-            {
-                continue;
-            }
-
-            // only metagems to be (de)activated, so only enchants with condition
-            uint32 condition = enchantEntry->EnchantmentCondition;
-            if (condition)
-            {
-                ApplyEnchantment(pItem, EnchantmentSlot(enchant_slot), apply);
-            }
-        }
     }
 }
